Adjust system mode behavior for Acova (Zehnder) heaters (#256)
* Adjust system mode behavior for Acova (Zehnder) heaters

This commit addresses an issue where SystemMode.AUTO for Acova (Zehnder)
heaters behaves differently from the HA standard. For this manufacturer,
SystemMode.AUTO allows control of the heater via Zigbee, while
SystemMode.HEAT puts the heater in manual mode.

As a result, the conversion tables in ZHA climate constants (HVACMode)
are incorrect for this device. The mapping for SystemMode.AUTO has been
updated to correspond to HVACMode.HEAT instead of HVACMode.HEAT_COOL.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>
Co-authored-by: Matthéo PERELLE <mattheo.perelle@gmail.com>

* Add test for Acova (Zehnder) thermostat

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>
Co-authored-by: Matthéo PERELLE <mattheo.perelle@gmail.com>

---------

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>
Co-authored-by: Matthéo PERELLE <mattheo.perelle@gmail.com>